refactor(comments): store times in milliseconds
The in-memory time field used whole seconds, so the player seek API
was typed as int and sub-second precision was impossible to express.
Milliseconds give an integer unit fine enough for frame-accurate
work that flows cleanly through the model and the editing UI, while
the seek command is retyped to a float so it can still receive the
converted value. No user-visible format change: documents are still
read and written as hh:mm:ss and all displayed output is identical.

+++ mpvqc/services/player/service.py
@@ -237,7 +237,7 @@ class PlayerService(QObject):
     def press_key(self, command: str) -> None:
         self._mpv_player.command_async("keypress", command)
 
-    def jump_to(self, seconds: int) -> None:
+    def jump_to(self, seconds: float) -> None:
         self._mpv_player.command_async("seek", seconds, "absolute+exact")
 
     def press_mouse_left(self) -> None:
